Refactor : 예외처리 전역으로 설정
