+++ src/main/java/com/sysone/ddogdog/common/exception/GlobalExceptionHandler.java
@@ -4,6 +4,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
+import org.springframework.web.servlet.NoHandlerFoundException;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
@@ -15,7 +16,17 @@ public class GlobalExceptionHandler {
         return new ResponseEntity<>(new ErrorResponse(errorCode), HttpStatus.NOT_FOUND);
     }
 
-    //공통
+    @ExceptionHandler(NoAvailableRoomsException.class)
+    public ResponseEntity<ErrorResponse> handleNoAvailableRoomsException(NoAvailableRoomsException ex) {
+        return new ResponseEntity<>(new ErrorResponse(CustomerErrorCode.NO_MORE_ROOMS), HttpStatus.NOT_FOUND);
+    }
+    @ExceptionHandler(NoHandlerFoundException.class)
+    public ResponseEntity<ErrorResponse> handleNotFoundException(NoHandlerFoundException ex) {
+        ErrorResponse errorResponse = new ErrorResponse(CommonErrorCode.NOT_FOUND);
+        return new ResponseEntity<>(errorResponse, HttpStatus.NOT_FOUND);
+    }
+
+  //  공통
     @ExceptionHandler(Exception.class)
     public ResponseEntity<ErrorResponse> handleGenericException(Exception ex) {
         return new ResponseEntity<>(new ErrorResponse(CommonErrorCode.INTERNAL_SERVER_ERROR),
