+++ src/main/java/com/sysone/ddogdog/customer/reservation/controller/ReservationController.java
@@ -5,7 +5,7 @@ import com.sysone.ddogdog.customer.reservation.model.RequestReservationDTO;
 import com.sysone.ddogdog.customer.reservation.model.ResponseMostReservationHotelDTO;
 import com.sysone.ddogdog.customer.reservation.model.ResponseReservationStatsDTO;
 import com.sysone.ddogdog.customer.reservation.service.ReservationService;
-import com.sysone.ddogdog.customer.roomChoice.exception.NoAvailableRoomsException;
+import com.sysone.ddogdog.common.exception.NoAvailableRoomsException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
@@ -24,12 +24,8 @@ public class ReservationController {
     @PostMapping
     public ResponseEntity<String> saveReserve(@AuthenticationPrincipal PrincipalDetails user,
         @RequestBody RequestReservationDTO dto) {
-        try {
             reservationService.saveReserve(user,dto);
             return new ResponseEntity<>("예약에 성공했습니다", HttpStatus.CREATED);
-        } catch (NoAvailableRoomsException e) {
-            return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_FOUND);
-        }
     }
 
     @PatchMapping("/{id}")
